Fix layout on safari
Previously I used `height: 100vh` on both the `html` and `body` tags. However
setting a `height` of a `100vh` broke the layout of pages where there was a lot
of content on Safari.

This commit fixes the issue by setting a `min-height: 100%` the `body` element
and setting a `height: 100%` on the `html` tag.

**NOTES**
+ Using `min-height` works fine on all the major browsers apart from IE, so I'll
  be adding a mix-in that's specific to IE to use the `height` instead.